requirements.txt
dj-database-url==0.5.0
Django==4.0.7
django-cors-headers==3.13.0
django-filter==2.4.0
djangorestframework==3.13.1
django-model-utils==4.0.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
python-dateutil==2.8.1
PyYAML==6.0
pytz==2021.3
bcrypt==3.1.7
djangorestframework-simplejwt==4.8.0
django-phonenumber-field[phonenumbers]==6.3.0
pyjwt[crypto]==2.4.0
coreapi==2.3.3
djangorestframework-api-key==2.0.0
